+++ src/main/java/com/swp/config/SecurityConfig.java
@@ -12,8 +12,7 @@ import com.swp.auth.JwtAuthenticationEntryPoint;
 import com.swp.auth.JwtAuthenticationFilter;
 import com.swp.auth.JwtProvider;
 import com.swp.oauth.CookieOAuth2AuthorizationRequestRepository;
-import com.swp.oauth.OAuth2FailureHandler;
-import com.swp.oauth.OAuth2SuccessHandler;
+import com.swp.oauth.OAuth2ResultHandler;
 import com.swp.oauth.ThirdPartyOAuth2UserService;
 import com.swp.user.domain.Role;
 
@@ -24,8 +23,7 @@ import lombok.RequiredArgsConstructor;
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
 	private final ThirdPartyOAuth2UserService thirdPartyOAuth2UserService;
-	private final OAuth2SuccessHandler successHandler;
-	private final OAuth2FailureHandler failureHandler;
+	private final OAuth2ResultHandler oAuth2ResultHandler;
 	private final JwtProvider jwtProvider;
 	private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
 	private final CookieOAuth2AuthorizationRequestRepository oAuth2AuthorizationRequestRepository;
@@ -65,8 +63,8 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 			.userInfoEndpoint()
 			.userService(thirdPartyOAuth2UserService)
 			.and()
-			.successHandler(successHandler)
-			.failureHandler(failureHandler);
+			.successHandler(oAuth2ResultHandler)
+			.failureHandler(oAuth2ResultHandler);
 
 		http.addFilterBefore(new JwtAuthenticationFilter(jwtProvider, jwtAuthenticationEntryPoint),
 			UsernamePasswordAuthenticationFilter.class);
