+++ src/main/java/com/swp/oauth/OAuth2ResultHandler.java
@@ -3,6 +3,7 @@ package com.swp.oauth;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.Optional;
+import java.util.regex.Pattern;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.Cookie;
@@ -11,7 +12,9 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -25,7 +28,7 @@ import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 @Component
-public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
+public class OAuth2ResultHandler implements AuthenticationSuccessHandler, AuthenticationFailureHandler {
 
 	private static final String PROVIDER_KEY = "provider";
 	@Value("${frontend.redirectUri}")
@@ -44,6 +47,24 @@ public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
 		response.sendRedirect(targetUrl);
 	}
 
+	@Override
+	public void onAuthenticationFailure(HttpServletRequest request, HttpServletResponse response,
+		AuthenticationException exception) throws IOException, ServletException {
+		String targetUrl = getTargetUrl(request);
+		requestRepository.removeAuthorizationRequestCookies(request, response);
+		response.sendRedirect(targetUrl);
+	}
+
+	private String getTargetUrl(HttpServletRequest request) {
+		Optional<String> redirectUri = CookieUtils.getCookie(request,
+			CookieOAuth2AuthorizationRequestRepository.REDIRECT_URI_PARAM_COOKIE_NAME).map(Cookie::getValue);
+
+		if (redirectUri.isPresent() && isAuthorizedPattern(redirectUri)) {
+			return UriComponentsBuilder.fromUriString(redirectUri.get()).build().toUriString();
+		}
+		return REDIRECTION_URI;
+	}
+
 	private String getTargetUrl(HttpServletRequest request, OAuth2User oAuth2User) {
 		String provider = oAuth2User.getAttribute(PROVIDER_KEY);
 		String providerId = oAuth2User.getName();
@@ -54,8 +75,7 @@ public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
 			CookieOAuth2AuthorizationRequestRepository.REDIRECT_URI_PARAM_COOKIE_NAME).map(Cookie::getValue);
 		String redirectUri = REDIRECTION_URI;
 
-		if (cookieRedirectUri.isPresent() && cookieRedirectUri.stream()
-			.anyMatch(uri -> Arrays.asList(AUTHORIZED_URIS).contains(uri))) {
+		if (cookieRedirectUri.isPresent() && isAuthorizedPattern(cookieRedirectUri)) {
 			redirectUri = cookieRedirectUri.get();
 		}
 		return UriComponentsBuilder.fromUriString(redirectUri)
@@ -63,5 +83,11 @@ public class OAuth2SuccessHandler implements AuthenticationSuccessHandler {
 			.queryParam("refreshToken", token.getRefreshToken())
 			.build()
 			.toUriString();
+	}
+
+	private boolean isAuthorizedPattern(Optional<String> cookieRedirectUri) {
+		return cookieRedirectUri.stream()
+			.anyMatch(uri -> Arrays.stream(AUTHORIZED_URIS)
+				.anyMatch(authorizedUri -> Pattern.matches(authorizedUri, uri)));
 	}
 }
